Add missing functionality for frontend rendering.
Also correct logic errors.

The frontend rendering is what this extension is all about and had been drafted since almost a year now. In the simplest form this is just a few lines of code, but in reality as a proper extension and in the scope of typo3 it turned into pages.
The benefit is that it's now fully configurable and integrates tightly with the Typo3 backend.

Quick tutorial:

* Install extension. (manually into typo3conf/ext/ or via typo3 extension manager).
* Create a (system/menu-hidden) folder, e.g. RecentEntryCollections. Created and added an image for detailed instructions here in a great community-effort driven help resource: http://docs.typo3.org/typo3cms/ExtbaseFluidBook/4-FirstExtension/4-make-products-persistent.html
* Create a recent entry collection. (see image in the instructions referenced above).
* [optional: Create criteria for limiting the content elements to collect, e.g. pageToCollectFrom or typeToCollect. Follow the same procedure as for adding a new recentEntryCollection via  Enter Folder -> Add New -> CollectRecentEntries: e.g. typeToCollect .]
* Finally add the plugin to any page you like.

* Add more recent entry collections and combine criteria to get a nice overview page showing a mix of the most recent entries across the website.